+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Beta 0.5.1] 2025-02-13
+
+### Changed
+- clockinrecord's format
+
+---
 ## [Beta 0.5.0] 2025-02-13
 
 ### Added